Add 2 new variants to `RepositoryChangeStatus`
Add `Error`, which is used when getting the status of a repository
resulted in an error.

Add `Unknown`, which is used when the status is unknown.
Unknown is different from Error, as it doesn't suggest that there is or
isn't a problem.

Both Error and Unknown cover cases that would have previously been
ambiguous.