Add `list_repositories_with_uncommitted_changes`
This function will list all repositories, but filter out any uncommitted
changes. This function is useful, for when you want to display only the
repositories which require action. For example, when you want to make
sure that there are no repositories that may be lost if you lose access
to the data on your machine.